code for loading CIFAR-10

+++ seann/layers/std/DenseLayer.cuh
@@ -31,7 +31,7 @@ namespace seann {
             logInfo(LOG_SEG_SEANN, "Registered FC: Input:" + to_string(inputSize) +
                    " Output:" + to_string(outputSize));
             logDebug(LOG_SEG_SEANN,"Current total memory occupation : " +
-                to_string(MEMORY_OCCUPATION/(1024*1024)));
+                to_string((double)MEMORY_OCCUPATION/(1024*1024)));
         }
 
         //forward activation
